refactor(app): move and rename Repository
