+++ p-week-01/TodoApp/Repositories/ApplicationContext.cs
@@ -3,12 +3,13 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using TodoApp.Models;
 
-namespace TodoApp.Models
+namespace TodoApp.Repositories
 {
-    public class TodoContext : DbContext
+    public class ApplicationContext : DbContext
     {
-        public TodoContext(DbContextOptions<TodoContext> options) : base(options)
+        public ApplicationContext(DbContextOptions<ApplicationContext> options) : base(options)
         {
         }
